requirements.txt
fastapi
uvicorn
python-multipart
openai
openai-whisper